+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Raytyping Changelog
 
-## [Initial Version] - 2023-12-02
+## [Initial Version] - 2023-12-03
+
+- Initial version of Raytyping
+- Allow user to type in a text box
+- Show results of typing (WPM, accuracy)